[Feature][transform-v2] SQL transform support schema change
