+++ seatunnel-transforms-v2/src/main/java/org/apache/seatunnel/transform/sql/SQLTransform.java
@@ -17,6 +17,8 @@
 
 package org.apache.seatunnel.transform.sql;
 
+import org.apache.seatunnel.shade.com.google.common.annotations.VisibleForTesting;
+
 import org.apache.seatunnel.api.common.CommonOptions;
 import org.apache.seatunnel.api.configuration.Option;
 import org.apache.seatunnel.api.configuration.Options;
@@ -27,6 +29,15 @@ import org.apache.seatunnel.api.table.catalog.ConstraintKey;
 import org.apache.seatunnel.api.table.catalog.PhysicalColumn;
 import org.apache.seatunnel.api.table.catalog.TableIdentifier;
 import org.apache.seatunnel.api.table.catalog.TableSchema;
+import org.apache.seatunnel.api.table.schema.event.AlterTableAddColumnEvent;
+import org.apache.seatunnel.api.table.schema.event.AlterTableChangeColumnEvent;
+import org.apache.seatunnel.api.table.schema.event.AlterTableColumnEvent;
+import org.apache.seatunnel.api.table.schema.event.AlterTableColumnsEvent;
+import org.apache.seatunnel.api.table.schema.event.AlterTableDropColumnEvent;
+import org.apache.seatunnel.api.table.schema.event.AlterTableModifyColumnEvent;
+import org.apache.seatunnel.api.table.schema.event.SchemaChangeEvent;
+import org.apache.seatunnel.api.table.schema.handler.TableSchemaChangeEventDispatcher;
+import org.apache.seatunnel.api.table.schema.handler.TableSchemaChangeEventHandler;
 import org.apache.seatunnel.api.table.type.SeaTunnelDataType;
 import org.apache.seatunnel.api.table.type.SeaTunnelRow;
 import org.apache.seatunnel.api.table.type.SeaTunnelRowType;
@@ -46,6 +57,7 @@ import static org.apache.seatunnel.transform.sql.SQLEngineFactory.EngineType.ZET
 @Slf4j
 public class SQLTransform extends AbstractCatalogSupportFlatMapTransform {
     public static final String PLUGIN_NAME = "Sql";
+    private final TableSchemaChangeEventHandler tableSchemaChangeEventHandler;
 
     public static final Option<String> KEY_QUERY =
             Options.key("query").stringType().noDefaultValue().withDescription("The query SQL");
@@ -67,7 +79,8 @@ public class SQLTransform extends AbstractCatalogSupportFlatMapTransform {
     private final String inputTableName;
 
     public SQLTransform(@NonNull ReadonlyConfig config, @NonNull CatalogTable catalogTable) {
-        super(config, catalogTable);
+        super(catalogTable);
+        this.tableSchemaChangeEventHandler = new TableSchemaChangeEventDispatcher();
         this.query = config.get(KEY_QUERY);
         if (config.getOptional(KEY_ENGINE).isPresent()) {
             this.engineType = EngineType.valueOf(config.get(KEY_ENGINE).toUpperCase());
@@ -111,6 +124,10 @@ public class SQLTransform extends AbstractCatalogSupportFlatMapTransform {
     }
 
     @Override
+    protected TableSchema transformTableSchema() {
+        return convertTableSchema(inputCatalogTable.getTableSchema());
+    }
+
     protected TableSchema convertTableSchema(TableSchema tableSchema) {
         tryOpen();
         List<String> inputColumnsMapping = new ArrayList<>();
@@ -173,6 +190,127 @@ public class SQLTransform extends AbstractCatalogSupportFlatMapTransform {
             columns.add(column);
         }
         return builder.columns(columns).build();
+    }
+
+    @Override
+    public SchemaChangeEvent mapSchemaChangeEvent(SchemaChangeEvent event) {
+
+        TableSchema newTableSchema =
+                tableSchemaChangeEventHandler
+                        .reset(inputCatalogTable.getTableSchema())
+                        .apply(event);
+        this.inputCatalogTable =
+                CatalogTable.of(
+                        inputCatalogTable.getTableId(),
+                        newTableSchema,
+                        inputCatalogTable.getOptions(),
+                        inputCatalogTable.getPartitionKeys(),
+                        inputCatalogTable.getComment());
+
+        if (event instanceof AlterTableColumnsEvent) {
+            AlterTableColumnsEvent alterTableColumnsEvent = (AlterTableColumnsEvent) event;
+            AlterTableColumnsEvent newEvent =
+                    new AlterTableColumnsEvent(
+                            event.tableIdentifier(),
+                            alterTableColumnsEvent.getEvents().stream()
+                                    .map(this::convertName)
+                                    .collect(Collectors.toList()));
+
+            newEvent.setJobId(event.getJobId());
+            newEvent.setStatement(((AlterTableColumnsEvent) event).getStatement());
+            newEvent.setSourceDialectName(((AlterTableColumnsEvent) event).getSourceDialectName());
+            if (event.getChangeAfter() != null) {
+                newEvent.setChangeAfter(
+                        CatalogTable.of(
+                                event.getChangeAfter().getTableId(), event.getChangeAfter()));
+            }
+            return newEvent;
+        }
+        if (event instanceof AlterTableColumnEvent) {
+            return convertName((AlterTableColumnEvent) event);
+        }
+        return event;
+    }
+
+    @VisibleForTesting
+    public AlterTableColumnEvent convertName(AlterTableColumnEvent event) {
+        AlterTableColumnEvent newEvent = event;
+        switch (event.getEventType()) {
+            case SCHEMA_CHANGE_ADD_COLUMN:
+                AlterTableAddColumnEvent addColumnEvent = (AlterTableAddColumnEvent) event;
+                newEvent =
+                        new AlterTableAddColumnEvent(
+                                event.tableIdentifier(),
+                                convertName(addColumnEvent.getColumn()),
+                                addColumnEvent.isFirst(),
+                                convertName(addColumnEvent.getAfterColumn()));
+                break;
+            case SCHEMA_CHANGE_DROP_COLUMN:
+                AlterTableDropColumnEvent dropColumnEvent = (AlterTableDropColumnEvent) event;
+                newEvent =
+                        new AlterTableDropColumnEvent(
+                                event.tableIdentifier(), convertName(dropColumnEvent.getColumn()));
+                break;
+            case SCHEMA_CHANGE_MODIFY_COLUMN:
+                AlterTableModifyColumnEvent modifyColumnEvent = (AlterTableModifyColumnEvent) event;
+                newEvent =
+                        new AlterTableModifyColumnEvent(
+                                event.tableIdentifier(),
+                                convertName(modifyColumnEvent.getColumn()),
+                                modifyColumnEvent.isFirst(),
+                                convertName(modifyColumnEvent.getAfterColumn()));
+                break;
+            case SCHEMA_CHANGE_CHANGE_COLUMN:
+                AlterTableChangeColumnEvent changeColumnEvent = (AlterTableChangeColumnEvent) event;
+                boolean nameChanged =
+                        !changeColumnEvent
+                                .getOldColumn()
+                                .equals(changeColumnEvent.getColumn().getName());
+                if (nameChanged) {
+                    log.warn(
+                            "FieldRenameTransform does not support changing column name, "
+                                    + "old column name: {}, new column name: {}",
+                            changeColumnEvent.getOldColumn(),
+                            changeColumnEvent.getColumn().getName());
+                    return changeColumnEvent;
+                }
+
+                newEvent =
+                        new AlterTableChangeColumnEvent(
+                                event.tableIdentifier(),
+                                convertName(changeColumnEvent.getOldColumn()),
+                                convertName(changeColumnEvent.getColumn()),
+                                changeColumnEvent.isFirst(),
+                                convertName(changeColumnEvent.getAfterColumn()));
+                break;
+            default:
+                log.warn("Unsupported event: {}", event);
+                return event;
+        }
+
+        newEvent.setJobId(event.getJobId());
+        newEvent.setStatement(event.getStatement());
+        newEvent.setSourceDialectName(event.getSourceDialectName());
+        if (event.getChangeAfter() != null) {
+            CatalogTable newChangeAfter =
+                    CatalogTable.of(
+                            event.getChangeAfter().getTableId(),
+                            convertTableSchema(event.getChangeAfter().getTableSchema()),
+                            event.getChangeAfter().getOptions(),
+                            event.getChangeAfter().getPartitionKeys(),
+                            event.getChangeAfter().getComment());
+            newEvent.setChangeAfter(newChangeAfter);
+        }
+        return newEvent;
+    }
+
+    @VisibleForTesting
+    public String convertName(String name) {
+        return name;
+    }
+
+    private Column convertName(Column column) {
+        return column.rename(convertName(column.getName()));
     }
 
     @Override
